fix: refactoring code to input url
