+++ cmd/datetimecli/main.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"fmt"
+	// "os"
+	"github.com/codescalersinternships/DateTime-HTTP-Client-Rowan/pkg"
+)
+
+func main() {
+	client := pkg.NewDateTimeClient("http://localhost:8080")
+	data, _ := client.GetDateTime()
+	fmt.Printf("%s", data)
+}
