+++ pkg/client_test.go
@@ -1,0 +1,5 @@
+package pkg
+
+import (
+// "tests"
+)
